Moves mock data to a separate file
