+++ js/form.js
@@ -3,7 +3,7 @@ import { resetScale } from './scale.js';
 import { resetToDefault } from './effects.js';
 import { isValid, resetValidation } from './validation.js';
 import { sendData } from './api.js';
-import { showSuccessMessage, showUploadErrorMessage } from './form-messages.js';
+import { showSuccessMessage, showUploadErrorMessage } from './status-messages.js';
 import { ErrorText, SubmitButtonStatus } from './constants.js';
 
 const imageUploadForm = document.querySelector('.img-upload__form');
