+++ js/main.js
@@ -3,13 +3,11 @@ import './modal-window.js';
 import './validation.js';
 import './scale.js';
 import './effects.js';
-// import { PHOTOS_COUNT } from './constants.js';
-// import { getPhotos } from './data.js';
 import { getPicturePreview } from './previews.js';
 import { getData } from './api.js';
 import { setFormSubmit, closeImageUploadForm } from './form.js';
 import { ErrorText } from './constants';
-import { showErrorMessage } from './loading-error.js';
+import { showErrorMessage } from './status-messages.js';
 
 getData()
   .then((pictures) => {
