+++ js/status-messages.js
@@ -1,5 +1,6 @@
 import { isEscKey } from './utils.js';
 import { onFormEscKeydown } from './form.js';
+import { TIMEOUT } from './constants.js';
 
 const uploadErrorMessageTemplate = document.querySelector('#error').content.querySelector('.error');
 const uploadErrorMessageElement = uploadErrorMessageTemplate.cloneNode('true');
@@ -9,6 +10,21 @@ const uploadSuccessMessageTemplate = document.querySelector('#success').content.
 const uploadSuccessMessageElement = uploadSuccessMessageTemplate.cloneNode('true');
 const uploadSuccessMessageText = uploadSuccessMessageTemplate.querySelector('.success__title');
 
+const dataDownloadErrorTemplate = document.querySelector('#data-error').content.querySelector('.data-error');
+const dataDownloadErrorElement = dataDownloadErrorTemplate.cloneNode('true');
+const dataDownloadErrorMessage = dataDownloadErrorTemplate.querySelector('.data-error__title');
+
+
+const removeErrorMessage = () => document.body.remove(dataDownloadErrorElement);
+
+const closeErrorMessage = () => setTimeout(removeErrorMessage, TIMEOUT);
+
+const showErrorMessage = (message) => {
+  document.body.appendChild(dataDownloadErrorElement);
+  dataDownloadErrorMessage.textContent = message;
+  closeErrorMessage();
+};
+
 const hideMessage = () => {
   document.removeEventListener('keydown', onDocumentKeydown);
   document.addEventListener('keydown', onFormEscKeydown);
@@ -17,8 +33,8 @@ const hideMessage = () => {
 };
 
 const onSuccessButtonClick = () => {
-  document.removeEventListener('keydown', onFormEscKeydown);
   hideMessage();
+  document.removeEventListener('keydown', onFormEscKeydown);
 };
 
 const showSuccessMessage = (message) => {
@@ -26,12 +42,6 @@ const showSuccessMessage = (message) => {
   uploadSuccessMessageText.textContent = message;
   document.addEventListener('keydown', onDocumentKeydown);
 };
-
-uploadSuccessMessageElement.addEventListener('click', (evt) => {
-  if (evt.target.classList.contains('success__button') || evt.target.classList.contains('success')) {
-    onSuccessButtonClick();
-  }
-});
 
 const onErrorButtonClick = () => hideMessage();
 
@@ -41,6 +51,12 @@ const showUploadErrorMessage = (message) => {
   document.removeEventListener('keydown', onFormEscKeydown);
   document.addEventListener('keydown', onDocumentKeydown);
 };
+
+uploadSuccessMessageElement.addEventListener('click', (evt) => {
+  if (evt.target.classList.contains('success__button') || evt.target.classList.contains('success')) {
+    onSuccessButtonClick();
+  }
+});
 
 uploadErrorMessageElement.addEventListener('click', (evt) => {
   if (evt.target.classList.contains('error__button') || evt.target.classList.contains('error')) {
@@ -55,4 +71,4 @@ function onDocumentKeydown(evt) {
   }
 }
 
-export { showUploadErrorMessage, showSuccessMessage };
+export { showErrorMessage, showUploadErrorMessage, showSuccessMessage };
